+++ src/app/core/components/history.component.ts
@@ -1,30 +1,23 @@
-import { Component, OnInit } from "@angular/core";
+import { Component, OnInit, OnDestroy } from "@angular/core";
+import { ActivatedRoute, ParamMap } from "@angular/router";
 import { FormControl } from "@angular/forms";
-import { MatDialog } from "@angular/material";
-import { Project, AuthService, ProjectService, Element, ElementItem, ElementField, ElementCell, NotificationService, ElementFieldDataType } from "@forcrowd/backbone-client-core";
-import { finalize } from "rxjs/operators";
-
-// Component
-import { RemoveHistoryConfirmComponent } from "./remove-history.component";
-import { ConfirmEditComponent } from "./confirm-edit.component";
-
-// Service
-import { AppProjectService } from "../app-core.module";
-import { ProfileRemoveProjectComponent } from './profile-remove-project.component';
+import { Project, ProjectService, Element, ElementField, ElementCell, AuthService } from "@forcrowd/backbone-client-core";
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: "history",
   templateUrl: "history.component.html",
   styleUrls: ["history.component.css"],
 })
-export class HistoryComponent implements OnInit {
+export class HistoryComponent implements OnInit, OnDestroy {
 
-  entry: string = "";
-  changeElementName: boolean = false;
-  isBusy: boolean;
   project: Project = null;
   selectedTab = new FormControl(0);
   selectedTimeline: number = 0;
+  username: string = null;
+  timeline: string = null;
+  isBusy: boolean;
+  subscriptions: Subscription[] = [];
 
   get currentUser() {
     return this.authService.currentUser;
@@ -90,214 +83,40 @@ export class HistoryComponent implements OnInit {
     }
 
   constructor(private authService: AuthService,
-    private projectService: ProjectService,
-    private notificationService: NotificationService,
-    private dialog: MatDialog) {
-  }
-
-  cancelEditing() {
-    this.entry = "";
-    this.selectedElementCell = null;
-    this.isBusy = false;
-    this.changeElementName = false;
-  }
-
-  // Create project (only one time)
-  createProjectHistory(): void {
-    if (this.project === null) {
-      this.project = (this.projectService as AppProjectService).createProjectHistory();
-      this.projectService.saveChanges().subscribe(() => {
-        this.isBusy = false;
+    private activatedRoute: ActivatedRoute,
+    private projectService: ProjectService) {
+    if (activatedRoute.firstChild !== null) {
+      var activatedRouteSubscription = activatedRoute.firstChild.paramMap.subscribe((params: ParamMap): void => {
+        this.timeline = params.get("timeline");
+        if (this.project !== null) this.loadProject(this.project.Id);
       });
+      this.subscriptions.push(activatedRouteSubscription);
     }
   }
 
-  // Create a new entry in timeline
-  createHistroyContent(): void {
-
-    if (!this.currentUser || !this.currentUser.isAuthenticated()) return;
-
-    this.isBusy = true;
-
-    if (!this.changeElementName) {
-      if (this.selectedElementCell === null) {
-
-        // New Item
-        const elementItem = this.projectService.createElementItem({
-          Element: this.selectedElement,
-          Name: `Entry ${this.selectedElement.ElementItemSet.length + 1}`,
-        }) as ElementItem;
-
-        // Cell
-        this.projectService.createElementCell({
-          ElementField: this.selectedElementField,
-          ElementItem: elementItem,
-          StringValue: this.entry,
-        });
-
-        /* Like - Dislike */
-
-        // Item
-        const likeItem = this.projectService.createElementItem({
-          Element: this.selectedElement,
-          Name: `likes ${this.selectedElement.ElementFieldSet[1].ElementCellSet.length + 1}`
-        }) as ElementItem;
-        console.log("field ", this.selectedElement.ElementFieldSet[1]);
-
-        // Cell
-        const likeCell = this.projectService.createElementCell({
-          ElementField: this.selectedElement.ElementFieldSet[1],
-          ElementItem: likeItem
-        });
-
-        this.projectService.createUserElementCell(likeCell, 0);
-
-        this.projectService.saveChanges().subscribe(() => {
-          this.entry = "";
-          this.loadProject(this.project.Id);
-          this.isBusy = false;
-        });
-
-      } else {
-        this.selectedElementCell.StringValue !== this.entry ? this.change()
-          : this.isBusy = false;
-      }
-    } else {
-      // if timeline name will be change then
-      this.selectedElement.Name = this.entry;
-      this.projectService.saveChanges().subscribe(() => {
-        this.entry = "";
-        this.isBusy = false;
-      });
-      this.changeElementName = false;
-    }
-
+  decimalValueTotal(index: number): number {
+    return this.selectedElement.ElementFieldSet[1].ElementCellSet[index].DecimalValueTotal;
   }
 
-  // Create a new History Timeline
-  createNewHistroy(): void {
-    this.isBusy = true;
+  changeLikeCount(value: number, index: number): void {
+    const userElementCellSet = this.selectedElement.ElementFieldSet[1].ElementCellSet[index];
 
-    // New Timeline Element
-    const element = this.projectService.createElement({
-      Project: this.project,
-      Name: this.entry
-    }) as Element;
+    userElementCellSet.UserElementCellSet.length > 0 ? userElementCellSet.UserElementCellSet[0].DecimalValue = value
+      : this.projectService.createUserElementCell(userElementCellSet, value);
 
-    // Field
-    const elementField = this.projectService.createElementField({
-      Element: element,
-      Name: this.entry,
-      DataType: ElementFieldDataType.String,
-      SortOrder: 1
-    }) as ElementField;
-
-    // Item
-    const elementItem = this.projectService.createElementItem({
-      Element: element,
-      Name: this.entry
-    }) as ElementItem;
-
-    // Cell
-    this.projectService.createElementCell({
-      ElementField: elementField,
-      ElementItem: elementItem,
-      StringValue: "First",
-    });
-
-    /* --- */
-
-    // Like Dislikes
-    const elementField2 = this.projectService.createElementField({
-      Element: element,
-      Name: "likes",
-      DataType: ElementFieldDataType.Decimal,
-      UseFixedValue: false,
-      RatingEnabled: false,
-      SortOrder: 0
-    }) as ElementField;
-
-    // Item
-    const elementItem2 = this.projectService.createElementItem({
-      Element: element,
-      Name: "likes"
-    }) as ElementItem;
-
-    // Cell
-    const cell2 = this.projectService.createElementCell({
-      ElementField: elementField2,
-      ElementItem: elementItem2
-    });
-
-    this.projectService.createUserElementCell(cell2, 0);
-
-    this.projectService.saveChanges().subscribe(() => {
-      this.selectedTab.setValue(this.project.ElementSet.length + 1);
-      this.entry = "";
-      this.isBusy = false;
-    });
+    this.projectService.saveChanges().subscribe();
   }
 
-  // Change timeline item value
-  change(): void {
-    // change cell string value
-    const dialogRef = this.dialog.open(ConfirmEditComponent);
-    dialogRef.afterClosed().subscribe(confirmed => {
-
-      if (!confirmed) {
-        this.cancelEditing();
-        return;
-      }
-
-      this.selectedElementCell.StringValue = this.entry;
-      this.projectService.saveChanges().subscribe(() => {
-        this.notificationService.notification.next("Timeline item has been change");
-        this.entry = "";
-        this.selectedElementCell = null;
-        this.loadProject(this.project.Id);
-        this.isBusy = false;
-      });
-    });
-
+  // Set selected timeline element
+  selectTimeline(value: number): void {
+    this.selectedTimeline = value;
+    this.selectedElement = this.project.ElementSet[value];
+    this.selectedElementField = this.selectedElement.ElementFieldSet[0];
+    this.selectedElementLikeCountSet = this.selectedElement.ElementFieldSet[1].ElementCellSet; // for Like - Dislike
+    this.selectedElementCellSet = this.selectedElementField.ElementCellSet as ElementCell[];
+    this.selectedElementCellSet = this.selectedElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
   }
 
-  // Delete Timeline (history)
-  deleteTimeline(element: Element): void {
-
-    const dialogRef = this.dialog.open(ProfileRemoveProjectComponent);
-
-    dialogRef.afterClosed().subscribe(confirmed => {
-
-      if (!confirmed) {
-        return;
-      }
-
-      this.projectService.removeElement(element);
-      this.projectService.saveChanges().pipe(
-        finalize(() => {
-          if (this.project.ElementSet.length > 0) {
-            this.selectTimeline(0);
-            this.selectedTab.setValue(0);
-          }
-          this.notificationService.notification.next("Your timeline has been removed!");
-        })).subscribe();
-    });
-  }
-
-  // Edit Timeline Header (Chane element name)
-  editTimelineHeader(element: Element): void {
-    this.notificationService.notification.next("Please write new timeline name form to input then submit");
-    this.changeElementName = true;
-    this.entry = this.selectedElement.Name;
-  }
-
-  // Edit item
-  edit(elementCell: ElementCell) {
-    this.entry = elementCell.StringValue;
-    this.selectedElementCell = elementCell;
-  }
-
-  // Set project element and field
   loadProject(projectId: number) {
     this.projectService.getProjectExpanded<Project>(projectId)
       .subscribe(project => {
@@ -309,88 +128,50 @@ export class HistoryComponent implements OnInit {
         // Project History
         this.project = project;
 
-        if (this.project.ElementSet.length > 0) {
-          // Selected element
+        if (this.timeline !== null) {
+          for (var i = 0; i < this.project.ElementSet.length; i++) {
+            if (this.timeline === this.project.ElementSet[i].Name) {
+              this.selectTimeline(i);
+              return;
+            }
+          }
+        } else {
           this.selectedElement = this.project.ElementSet[this.selectedTimeline];
-
-          // ElementField
-          this.selectedElementField = this.selectedElement.ElementFieldSet[0];
-
-          // for Like - Dislike
-          // ElementLikeCountSet (ElementCellSet)
-          this.selectedElementLikeCountSet = this.selectedElement.ElementFieldSet[1].ElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
-
-          // ElementCellSet
-          this.selectedElementCellSet = this.selectedElementField.ElementCellSet as ElementCell[];
-          this.selectedElementCellSet = this.selectedElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
         }
 
-        this.isBusy = false;
+        this.selectedElementField = this.selectedElement.ElementFieldSet[0];
+        this.selectedElementCellSet = this.selectedElementField.ElementCellSet as ElementCell[];
+        this.selectedElementLikeCountSet = this.selectedElement.ElementFieldSet[1].ElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
+        this.selectedElementCellSet = this.selectedElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
       });
   }
 
-  changeLikeCount(value: number, index: number): void {
-    this.isBusy = true;
-    this.selectedElementLikeCountSet[index].UserElementCellSet[0].DecimalValue += value;
-    this.projectService.saveChanges().subscribe(() => {
-      this.isBusy = false;
-    });
-  }
-
-  // Set selected timeline element
-  selectTimeline(value: number): void {
-    this.isBusy = true;
-    this.selectedTimeline = value;
-    this.selectedElement = this.project.ElementSet[value];
-    this.selectedElementField = this.selectedElement.ElementFieldSet[0];
-    this.selectedElementLikeCountSet = this.selectedElement.ElementFieldSet[1].ElementCellSet; // for Like - Dislike
-    this.selectedElementCellSet = this.selectedElementField.ElementCellSet as ElementCell[];
-    this.selectedElementCellSet = this.selectedElementCellSet.sort((a, b) => (b.CreatedOn.getTime() - a.CreatedOn.getTime()));
-    this.isBusy = false;
-  }
-
-  // Remove selected item
-  removeHistoryItem(elementItem: ElementItem) {
-    const dialogRef = this.dialog.open(RemoveHistoryConfirmComponent);
-
-    dialogRef.afterClosed().subscribe(confirmed => {
-
-      if (!confirmed) return;
-
-      var likeItemIndex =  this.selectedElement.ElementItemSet.indexOf(elementItem);
-
-      this.projectService.removeElementItem(elementItem);
-      this.projectService.removeElementItem(this.selectedElement.ElementItemSet[likeItemIndex]);
-
-      this.projectService.saveChanges().pipe(
-        finalize(() => {
-          this.entry = "";
-          this.selectedElementCell = null;
-          this.loadProject(this.project.Id);
-        })).subscribe();
-    });
-
-  }
-
   ngOnInit(): void {
-    this.isBusy = true;
+    if (this.currentUser.isAuthenticated !== null) this.authService.ensureAuthenticatedUser().subscribe();
 
-    if (!this.currentUser || !this.currentUser.isAuthenticated()) return;
+    this.username = this.activatedRoute.snapshot.params["username"];
 
-    this.authService.getUser(this.currentUser.UserName).subscribe(() => {
+    this.projectService.getProjectSet<Project>(this.username)
+      .subscribe(projectSet => {
 
-      for (var i = 0; i < this.currentUser.ProjectSet.length; i++) {
-        var project = this.currentUser.ProjectSet[i];
-        if (project.Name === "History App") {
-          this.project = project;
-          this.loadProject(this.project.Id);
-          break;
-        }
-      }
+        for (var i = 0; i < projectSet.length; i++) {
+          var project = projectSet[i];
 
-      if (this.project === null) this.createProjectHistory();
-    });
+          if (project.Name === "History App") {
+            // Project History
+            this.project = project;
+            this.loadProject(this.project.Id);
+            return;
+          }
 
+        };
+      });
+  }
+
+  ngOnDestroy(): void {
+    for (let i = 0; i < this.subscriptions.length; i++) {
+      this.subscriptions[i].unsubscribe();
+    }
   }
 
 }